requirements.txt
keras==2.10.0
numpy==1.23.4
pillow==8.4.0
streamlit==1.18.1
streamlit_drawable_canvas==0.9.3
tensorflow==2.10.0
tensorflow_intel==2.10.0
